made changes as per comments

+++ egs/aspire/s5/local/multi_condition/normalize_wavs.py
@@ -31,7 +31,7 @@ def get_normalization_coefficient(file_list, is_rir, additional_scaling):
         assert(rate == sampling_rate)
       else:
         sampling_rate = rate
-      data = int(data/dtype_max_value)
+      data = data/dtype_max_value
       if is_rir:
         # just count the energy of the direct impulse response
         # this is treated as energy of signal from 0.001 seconds before impulse
@@ -57,7 +57,7 @@ def get_normalization_coefficient(file_list, is_rir, additional_scaling):
     except IOError:
       warnings.warn("Did not find the file {0}.".format(file))
   assert(total_samples > 0)
-  scaling_coefficient = np.sqrt(int(total_samples/total_energy))
+  scaling_coefficient = np.sqrt(total_samples/total_energy)
   print("Scaling coefficient is {0}.".format(scaling_coefficient))
   if math.isnan(scaling_coefficient):
     raise Exception(" Nan encountered while computing scaling coefficient. This is mostly due to numerical overflow")
