+++ egs/bentham/v1/local/gen_topo.py
@@ -69,7 +69,7 @@ print("<ForPhones>")
 print(" ".join([str(x) for x in silence_phones]))
 print("</ForPhones>")
 if(args.num_sil_states > 1):
-    transp = int(1.0/(args.num_sil_states - 1))
+    transp = 1.0/(args.num_sil_states - 1)
     
     state_str = "<State> 0 <PdfClass> 0 "
     for x in range(0, (args.num_sil_states - 1)):
