+++ egs/fisher_swbd/s5/local/format_acronyms_ctm_rt03.py
@@ -28,7 +28,7 @@ for line in fin:
     if items[4].find(".") != -1:
         letters = items[4].split("._")
         acronym_period = round(float(items[3]), 2)
-        letter_slot = round(int(acronym_period/ len(letters)), 2)
+        letter_slot = round(acronym_period/ len(letters), 2)
         time_start = round(float(items[2]), 2)
         for l in letters[:-1]:
             time = " %.2f %.2f " % (time_start, letter_slot)
